+++ examples/kdv_1d/kdv_1d_non_dim.jl
@@ -3,7 +3,7 @@ using DispersiveShallowWater
 using SummationByPartsOperators: upwind_operators, periodic_derivative_operator
 
 ###############################################################################
-# Semidiscretization of the KdV equation 
+# Semidiscretization of the KdV equation
 
 function initial_condition_non_dimensional(x, t, equations::KdVEquation1D, mesh)
     c = 1 / 3
@@ -45,7 +45,7 @@ tspan = (0.0, 50.0)
 ode = semidiscretize(semi, tspan)
 
 summary_callback = SummaryCallback()
-analysis_callback = AnalysisCallback(semi; interval = 100,
+analysis_callback = AnalysisCallback(semi; interval = 2000,
                                      extra_analysis_errors = (:conservation_error,),
                                      extra_analysis_integrals = (waterheight_total,
                                                                  waterheight, entropy))
@@ -56,4 +56,4 @@ sol = solve(ode, Tsit5(), abstol = 1e-8, reltol = 1e-8,
             save_everystep = false, callback = callbacks, saveat = saveat)
 
 # Plot the solution transformed back to non dimensional variables.
-# plot(semi => sol, conversion = prim2nondim, plot_initial = true)
+# plot(semi => sol, conversion = prim2nondim, plot_analytical = true)