+++ src/DispersiveShallowWater.jl
@@ -17,7 +17,7 @@ module DispersiveShallowWater
 
 using BandedMatrices: BandedMatrix
 using DelimitedFiles: readdlm
-using DiffEqBase: DiffEqBase, terminate!
+using DiffEqBase: DiffEqBase
 using FastBroadcast: @..
 using ForwardDiff: ForwardDiff
 using Interpolations: Interpolations, linear_interpolation